fix(console): prevent showing error msg when log in
